+++ yesod-core/src/Yesod/Core.hs
@@ -42,7 +42,9 @@ module Yesod.Core
       -- * Defaults
     , defaultErrorHandler
     , defaultYesodMiddleware
+    , defaultYesodMiddlewareNoAuthCheck
     , authorizationCheck
+    , dispatchAuthorizationCheck
       -- * Data types
     , AuthResult (..)
     , unauthorizedI