+++ yesod-core/test/RouteLeafCompile/NestedPattern.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE TypeFamilies #-}
+module NestedPattern where
+import Yesod.Core.RouteLeaf
+import YesodCoreTest.RouteLeaf.Foundation
+
+bad :: AuthDispatch OrgR -> Bool
+bad (DelegationR _) = True